Bay12 merge - 16.03 (#107)

+++ code/modules/persistence/graffiti.dm
@@ -38,27 +38,32 @@
 	. = ..(user)
 	to_chat(user,  "It reads \"[message]\".")
 
-/obj/effect/decal/writing/attackby(obj/item/thing, mob/user)
-	if(isWelder(thing))
-		var/obj/item/weldingtool/welder = thing
-		if(welder.isOn() && welder.remove_fuel(0,user) && do_after(user, 0.5 SECONDS, src, DO_PUBLIC_UNIQUE) && !QDELETED(src))
-			playsound(src.loc, 'sound/items/Welder2.ogg', 50, 1)
-			user.visible_message(SPAN_NOTICE("\The [user] clears away some graffiti."))
-			qdel(src)
-	else if(thing.sharp)
 
-		if(jobban_isbanned(user, "Graffiti"))
-			to_chat(user, SPAN_WARNING("You are banned from leaving persistent information across rounds."))
-			return
+/obj/effect/decal/writing/use_tool(obj/item/tool, mob/user, list/click_params)
+	// Sharp Item - Engrave additional message
+	if (is_sharp(tool))
+		var/turf/target = get_turf(src)
+		target.try_graffiti(user, tool)
+		return TRUE
 
-		var/_message = sanitize(input("Enter an additional message to engrave.", "Graffiti") as null|text, trim = TRUE)
-		if(_message && loc && user && !user.incapacitated() && user.Adjacent(loc) && thing.loc == user)
-			user.visible_message(SPAN_WARNING("\The [user] begins carving something into \the [loc]."))
-			if(do_after(user, max(2 SECONDS, length(_message)), src, DO_PUBLIC_UNIQUE) && loc)
-				user.visible_message(SPAN_DANGER("\The [user] carves some graffiti into \the [loc]."))
-				message = "[message] [_message]"
-				author = user.ckey
-				if(lowertext(message) == "elbereth")
-					to_chat(user, SPAN_NOTICE("You feel much safer."))
-	else
-		. = ..()
+	// Welding Torch - Remove decal
+	if (isWelder(tool))
+		var/obj/item/weldingtool/welder = tool
+		if (!welder.can_use(1, user, "to remove \the [src]."))
+			return TRUE
+		playsound(src, 'sound/items/Welder2.ogg', 50, TRUE)
+		user.visible_message(
+			SPAN_NOTICE("\The [user] starts burning away \the [src] with \a [tool]."),
+			SPAN_NOTICE("You start burning away \the [src] with \the [tool].")
+		)
+		if (!do_after(user, 0.5 SECONDS, src, DO_REPAIR_CONSTRUCT) || !user.use_sanity_check(src, tool) || !welder.can_use(1, user, "to remove \the [src]"))
+			return TRUE
+		user.visible_message(
+			SPAN_NOTICE("\The [user] clears away \the [src] with \a [tool]."),
+			SPAN_NOTICE("You clear away \the [src] with \the [tool].")
+		)
+		playsound(src, 'sound/items/Welder2.ogg', 50, TRUE)
+		qdel(src)
+		return TRUE
+
+	return ..()
